+++ src/main/java/org/example/loginapi/auth/service/AuthService.java
@@ -10,7 +10,7 @@ import org.example.loginapi.auth.entity.User;
 import org.example.loginapi.auth.repository.UserRepository;
 import org.example.loginapi.common.exception.CustomException;
 import org.example.loginapi.common.exception.ErrorCode;
-import org.example.loginapi.jwt.JwtUtil;
+import org.example.loginapi.common.security.jwt.JwtUtil;
 import org.springframework.stereotype.Service;
 
 @Service
